fix(events): refactor abstract classes
