+++ packit_service/worker/allowlist.py
@@ -47,7 +47,7 @@ from packit_service.worker.events import (
     TestingFarmResultsEvent,
     github,
 )
-from packit_service.worker.events.comment import CommitCommentEvent
+from packit_service.worker.events.abstract.comment import Commit as CommitCommentEvent
 from packit_service.worker.events.koji.base import Build as KojiBuildEvent
 from packit_service.worker.events.koji.base import BuildTag as KojiBuildTagEvent
 from packit_service.worker.helpers.build import CoprBuildJobHelper
