+++ packit_service/worker/events/copr.py
@@ -15,7 +15,7 @@ from packit_service.models import (
     ProjectEventModelType,
     SRPMBuildModel,
 )
-from packit_service.worker.events.abstract import Result
+from packit_service.worker.events.abstract.base import Result
 from packit_service.worker.events.enums import FedmsgTopic
 
 logger = getLogger(__name__)
