+++ packit_service/worker/events/github/issue.py
@@ -5,7 +5,7 @@ from typing import Optional
 
 from ogr.abstract import Comment as OgrComment
 
-from packit_service.worker.events.comment import AbstractIssueCommentEvent
+from packit_service.worker.events.abstract.comment import Issue as AbstractIssueCommentEvent
 from packit_service.worker.events.enums import IssueCommentAction
 from packit_service.worker.events.github.abstract import GithubEvent
