+++ packit_service/worker/events/gitlab/commit.py
@@ -1,8 +1,8 @@
 # Copyright Contributors to the Packit project.
 # SPDX-License-Identifier: MIT
 
-from packit_service.worker.events.comment import CommitCommentEvent
+from packit_service.worker.events.abstract.comment import Commit
 
 
-class Comment(CommitCommentEvent):
+class Comment(Commit):
     pass
