+++ packit_service/worker/events/openscanhub/abstract.py
@@ -13,7 +13,7 @@ from packit_service.models import (
     OSHScanModel,
     ProjectEventModel,
 )
-from packit_service.worker.events.abstract import Result
+from packit_service.worker.events.abstract.base import Result
 
 logger = getLogger(__name__)
