+++ packit_service/worker/events/pagure/abstract.py
@@ -7,7 +7,7 @@ from typing import Optional
 from packit.config import PackageConfig
 
 from packit_service.config import PackageConfigGetter
-from packit_service.worker.events.abstract import ForgeIndependent
+from packit_service.worker.events.abstract.base import ForgeIndependent
 
 logger = getLogger(__name__)
