+++ packit_service/worker/handlers/copr.py
@@ -55,7 +55,7 @@ from packit_service.worker.events import (
     ReleaseEvent,
     ReleaseGitlabEvent,
 )
-from packit_service.worker.events.comment import CommitCommentEvent
+from packit_service.worker.events.abstract.comment import Commit as CommitCommentEvent
 from packit_service.worker.handlers.abstract import (
     JobHandler,
     RetriableJobHandler,
