+++ packit_service/worker/parser.py
@@ -62,7 +62,7 @@ from packit_service.worker.events import (
     VMImageBuildResultEvent,
     github,
 )
-from packit_service.worker.events.comment import CommitCommentEvent
+from packit_service.worker.events.abstract.comment import Commit as CommitCommentEvent
 from packit_service.worker.events.enums import (
     GitlabEventAction,
     IssueCommentAction,
