+++ tests/integration/test_pr_comment.py
@@ -67,7 +67,7 @@ from packit_service.utils import (
 from packit_service.worker.allowlist import Allowlist
 from packit_service.worker.celery_task import CeleryTask
 from packit_service.worker.events import PullRequestCommentPagureEvent
-from packit_service.worker.events.abstract import (
+from packit_service.worker.events.abstract.base import (
     ForgeIndependent as AbstractForgeIndependentEvent,
 )
 from packit_service.worker.handlers.bodhi import (
