+++ tests/unit/events/test_github.py
@@ -30,7 +30,7 @@ from packit_service.worker.events import (
     ReleaseEvent,
     github,
 )
-from packit_service.worker.events.comment import CommitCommentEvent
+from packit_service.worker.events.abstract.comment import Commit as CommitCommentEvent
 from packit_service.worker.events.enums import (
     IssueCommentAction,
     PullRequestAction,
